The most fondamental assertations are now made on first product

+++ app/Data/CarLoadInventory/CarLoadInventoryResultItemDTO.php
@@ -16,7 +16,7 @@ class CarLoadInventoryResultItemDTO
         public ConvertedQuantityDTO      $totalSoldConverted,
         public ConvertedQuantityDTO      $totalReturnedConverted,
         public ConvertedQuantityDTO $resultConverted,
-        public int $resultOfComputation = 0,
+        public float $resultOfComputation = 0,
         public int $priceOfResultComputation = 0
     )
     {
